feat: mac下加入执行start.sh打开浏览器的功能.
