+++ start.sh
@@ -1,1 +1,2 @@
-php -S 0.0.0.0:8000 -t web
+open -na "Google Chrome" --args --new-tab "http://127.0.0.1:9098/index.php"
+php -S 0.0.0.0:9098 -t web
